refactor(search): extract pure retrieveCandidates() core from SemanticSearch
Extracts the HNSW/BM25 retrieval + post-retrieval filtering (temporal/expiry/
supersede exclusion, scope.isAllowed() re-check) out of SemanticSearch.post()
into a side-effect-free retrieveCandidates() (resources/
semantic-retrieval-core.ts) that takes primitives only, never a Resource
instance. Auth resolution, rate-limiting, the reranker, and the
retrievalCount/lastRetrieved hit-tracking side effects stay in
SemanticSearch.post()'s wrapper, so an internal caller can reuse the core
retrieval logic without tripping them.

SemanticSearch's external behavior is unchanged: its full unit-test suite
stays green, and the isolated recall-harness (test/bench/recall-harness/)
returns byte-identical p@3/MRR (0.967/0.892 on the 87-record corpus,
hybrid=true) before and after this commit.
